# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CS 341 API",
    "description": "API for CS 341 Week 04",
    "version": "1.0.0"
  },
  "host": "safe-fjord-75598.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "Returns the default page",
        "parameters": [],
        "responses": {}
      }
    },
    "/static/css/{id}": {
      "get": {
        "description": "Returns any css files that are requested",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/static/js/{id}": {
      "get": {
        "description": "Returns any js files that are requested",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/manifest.json": {
      "get": {
        "description": "Returns the manifest file",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "description": "Returns user data if the token allows access (user ID matches and can read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "Updates a user's data if the token allows access (user ID matches and can delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "session": {
                  "example": "any"
                }
              }
            }
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "Deletes a user's data if the token allows access (user ID matches and can delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "session": {
                  "example": "any"
                }
              }
            }
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user": {
      "post": {
        "description": "Creates a new user from a username and password and returns a session token and ID of the inserted document",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "test@email.com"
                },
                "password": {
                  "example": "12345"
                },
                "firstName": {
                  "example": "Hunter"
                },
                "lastName": {
                  "example": "Livesay"
                },
                "favorites": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "ratings": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "embedding": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "number"
                  }
                },
                "styles": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/token": {
      "post": {
        "description": "Creates a new user from a google token and returns a session token and ID of the inserted document",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "test@email.com"
                },
                "password": {
                  "example": "12345"
                },
                "firstName": {
                  "example": "Hunter"
                },
                "lastName": {
                  "example": "Livesay"
                },
                "favorites": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "ratings": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "embedding": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "number"
                  }
                },
                "styles": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/authenticate": {
      "post": {
        "description": "Checks if a user exists and returns a session token if the password matches",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/{id}/favorites": {
      "get": {
        "description": "Returns a list of all the user's favorite outfits if the token allows access(user ID matches and can read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/{id}/ratings": {
      "get": {
        "description": "Returns a list of all the user's ratings if the token allows access(user ID matches and can read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/{id}/logout": {
      "delete": {
        "description": "Logs out a user by deleting any sessions related to them if the token allows access(user ID matches and can delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/{id}/embedding": {
      "get": {
        "description": "Returns a user's embedding if the token allows access(user ID matches and can read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/rating/{id}": {
      "get": {
        "description": "Gets a rating by id if the token allows access (userID matches rating createdByID and can read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/rating": {
      "post": {
        "description": "Adds a rating for a given user if the token allows access (userID matches and can write)",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "outfitId": {
                  "example": "any"
                },
                "styles": {
                  "example": [],
                  "type": "array",
                  "items": {
                    "type": "number"
                  }
                },
                "rating": {
                  "example": 1
                }
              }
            }
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/outfit": {
      "post": {
        "description": "Generates a new outfit based on the user's embedding and the styles they have selected, has a chance of generating a new outfit, instaed of selecting one from the database. ",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/outfit/{id}": {
      "get": {
        "description": "Returns an outfit by id if the token allows access (userID matches and can read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "session",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "Adds an outfit to a user's favorite list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "session": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/outfit/{id}/average": {
      "get": {
        "description": "Returns the average rating given to an outfit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}